cli, export types
