+++ packages/cli/.npmignore
@@ -1,7 +1,6 @@
-build/index*.*
-build/config*.*
-build/file_operations*.*
-build/*.tsbuildinfo
+build/*
+!build/bundle.js
+!build/types.d.ts
 src/*
 webpack.config.js
 tsconfig.json
