+++ packages/cli/src/config.ts
@@ -1,21 +1,6 @@
 import * as path from "path";
-import {ITranspilerOptions} from "@abaplint/transpiler";
 import * as fs from "fs";
-
-export interface ITranspilerConfig {
-  input_folder: string;
-  /** list of regex, case insensitive, empty gives all files, positive list */
-  input_filter?: string[];
-  output_folder: string;
-  /** to be deprecated */
-  lib?: string;
-  libs?: {
-    url: string,
-  }[],
-  write_unit_tests: boolean;
-  write_source_map: boolean;
-  options: ITranspilerOptions;
-}
+import {ITranspilerConfig} from "./types";
 
 export class TranspilerConfig {
