+++ performance/adhoc.ts
@@ -1,4 +1,4 @@
-import {ITranspilerConfig} from "../packages/cli/src/config";
+import {ITranspilerConfig} from "../packages/cli/src/types";
 import {FileOperations} from "../packages/cli/src/file_operations";
 import {compileFiles} from "test/_utils";
